requirements.txt
# Core Dependencies
streamlit==1.40.0
opencv-python==4.9.0.80
ultralytics>=8.3.121
deep-sort-realtime==1.3.2
wikipedia==1.4.0
google-genai>=0.4.0
python-dotenv==1.0.0

# Additional Dependencies - Python 3.12 Compatible
numpy<2,>=1.26.0
pillow>=11.0.0
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0